+++ scripts/build.js
@@ -155,7 +155,8 @@ async function buildReactFiles(copyright) {
     // Read and process files one at a time
     let filesChecked = 0;
     let filesUpdated = 0;
-    const allComponents = ["import React from 'react';"];
+    const defineExports = 'if (window.exports === undefined) { window.exports = window; }';
+    const allComponents = [];
     for (const component of components) {
         // Read both ES6 Class File and existing ES5 file
         const inFile = path.join(rootDir, 'es6', component + '.js');
@@ -167,10 +168,15 @@ async function buildReactFiles(copyright) {
         }
 
         // Convert ES6 code to ES5 using Babel
+        // Removing `import React` prevents Babel Standalone from requiring a
+        // custom `require` function and allows it to use the global `React` class
+        // which is already required by any custom `require` function.
+        codeES6 = codeES6.replace("import React from 'react';", '');
         let codeES5_New = Babel.transform(codeES6, options).code;
         if (isWindows) {
             codeES5_New = codeES5_New.replace(/\n/g, '\r\n');
         }
+        codeES5_New = codeES5_New.replace('Object.defineProperty(exports', defineExports + '\n\nObject.defineProperty(exports');
 
         // Compare - only update file if different
         if (codeES5_Old !== codeES5_New) {
@@ -179,9 +185,7 @@ async function buildReactFiles(copyright) {
             filesUpdated++;
         }
 
-        // Add code to array for the main [DataFormsJS.js] file, items such as "import React ..." must
-        // only be includes once so the text is modified as needed.
-        codeES6 = codeES6.replace("import React from 'react';", '');
+        // Add code to array for the main [DataFormsJS.js] file
         codeES6 = codeES6.replace('export default class', 'export class');
         codeES6 = codeES6.replace('@license', ''); // Required for all comments to be deleted
         allComponents.push(codeES6);
@@ -199,6 +203,7 @@ async function buildReactFiles(copyright) {
     `);
     let js = Babel.transform(allComponents.join('\n'), options).code;
     js = copyright + js;
+    js = js.replace('Object.defineProperty(exports', defineExports + '\n\nObject.defineProperty(exports');
 
     // Compare with existing [DataFormsJS.js] File.
     // When comparing convert CRLF -> LF to avoid issues.